Inject GitTrendAPI in to RepoListViewController
Removes the initialViewController from main storyboard and
the main Storyboard file from the project plist to allow the
`RepoListViewController` to be initialized manually within the
`SceneDelegate`.

This allows us to inject dependencies via the initializer and not
have to rely on property injection and implicitly unwrapped
optionals to inject non-optional values, immutable dependencies
like the `GitTrendAPI` for network calls.