Hide Status bar
There appears to be a known iOS 13 bug related
to status bar alignment for iPhone 8 and iPhoneSE
phones lacking the top notch present on new models.
This is likely related to the iOS 13 ViewController injection.

The existing shifting of the navigation bar text isn't sufficient
to make the BarButtonItems align as well.

For now the status bar will be disabled to enable
the refresh button. This will be fixed in a separate PR